chat_app/requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
psycopg2-binary
PyJWT
websockets
